abort info-sensitive commands in public channels
Louie accidentally destroyed a game by posting !newphase in a public channel and it showed a night action. We're not letting that happen again!

+++ commands/delete.js
@@ -13,6 +13,8 @@ module.exports = {
 		if (!gameState.players.length) 
 			return message.channel.send("There's nobody left to delete.\n\nYou monster...");
 
+		if (await Functions.CheckIfChannelVisible(message)) return;
+
 		if (!args.length)
 			return message.channel.send("Please include a player to delete from the game.");
